feat: move atlas link control to source study dataset list (#467)
