+++ app/components/Detail/components/ViewAtlasSourceDatasets/viewAtlasSourceDatasets.tsx
@@ -1,10 +1,8 @@
 import { COLLATOR_CASE_INSENSITIVE } from "@databiosphere/findable-ui/lib/common/constants";
 import { GridPaper } from "@databiosphere/findable-ui/lib/components/common/Paper/paper.styles";
 import { HCAAtlasTrackerSourceDataset } from "../../../../apis/catalog/hca-atlas-tracker/common/entities";
-import { PathParameter } from "../../../../common/entities";
 import { FormManager as FormManagerProps } from "../../../../hooks/useFormManager/common/entities";
 import { getAtlasSourceDatasetsTableColumns } from "../../../../viewModelBuilders/catalog/hca-atlas-tracker/common/viewModelBuilders";
-import { useSetLinkedAtlasSourceDatasets } from "../../../../views/AtlasSourceDatasetsView/hooks/useSetLinkedAtlasSourceDatasets";
 import { Paper } from "../../../Table/components/TablePaper/tablePaper.styles";
 import { TablePlaceholder } from "../../../Table/components/TablePlaceholder/tablePlaceholder";
 import { Table } from "../../../Table/table.styles";
@@ -14,41 +12,31 @@ import { TABLE_OPTIONS } from "./constants";
 interface ViewSourceDatasetsProps {
   atlasSourceDatasets?: HCAAtlasTrackerSourceDataset[];
   formManager: FormManagerProps;
-  pathParameter: PathParameter;
-  sourceStudiesSourceDatasets?: HCAAtlasTrackerSourceDataset[];
 }
 
 export const ViewAtlasSourceDatasets = ({
   atlasSourceDatasets = [],
   formManager,
-  pathParameter,
-  sourceStudiesSourceDatasets = [],
 }: ViewSourceDatasetsProps): JSX.Element => {
-  const linkedSourceDatasetIds = new Set(atlasSourceDatasets.map((d) => d.id));
   const {
     access: { canEdit, canView },
   } = formManager;
-  const { onSetLinked } = useSetLinkedAtlasSourceDatasets(pathParameter);
   if (!canView) return <RequestAccess />;
   return (
     <Paper>
       <GridPaper>
-        {sourceStudiesSourceDatasets.length > 0 && (
+        {atlasSourceDatasets.length > 0 && (
           <Table
-            columns={getAtlasSourceDatasetsTableColumns(
-              onSetLinked,
-              canEdit,
-              linkedSourceDatasetIds
-            )}
-            gridTemplateColumns="max-content auto auto"
-            items={sourceStudiesSourceDatasets.sort(sortSourceDataset)}
+            columns={getAtlasSourceDatasetsTableColumns()}
+            gridTemplateColumns="max-content auto"
+            items={atlasSourceDatasets.sort(sortSourceDataset)}
             tableOptions={TABLE_OPTIONS}
           />
         )}
         <TablePlaceholder
           canEdit={canEdit}
           message="No source datasets"
-          rowCount={sourceStudiesSourceDatasets.length}
+          rowCount={atlasSourceDatasets.length}
         />
       </GridPaper>
     </Paper>
