+++ app/views/SourceDatasetsView/sourceDatasetsView.tsx
@@ -9,6 +9,7 @@ import { Tabs } from "../../components/Detail/components/ViewSourceStudy/compone
 import { DetailView } from "../../components/Layout/components/Detail/detailView";
 import { useFetchAtlas } from "../../hooks/useFetchAtlas";
 import { useFormManager } from "../../hooks/useFormManager/useFormManager";
+import { useFetchAtlasSourceDatasets } from "../AtlasSourceDatasetsView/hooks/useFetchAtlasSourceDatasets";
 import { useFetchSourceStudy } from "../SourceStudyView/hooks/useFetchSourceStudy";
 import { getBreadcrumbs } from "./common/utils";
 import { useFetchSourceDatasets } from "./hooks/useFetchSourceDatasets";
@@ -23,6 +24,7 @@ export const SourceDatasetsView = ({
   const { atlas } = useFetchAtlas(pathParameter);
   const { sourceStudy } = useFetchSourceStudy(pathParameter);
   const { sourceDatasets } = useFetchSourceDatasets(pathParameter);
+  const { atlasSourceDatasets } = useFetchAtlasSourceDatasets(pathParameter);
   const formManager = useFormManager();
   const {
     access: { canEdit, canView },
@@ -41,6 +43,7 @@ export const SourceDatasetsView = ({
         }
         mainColumn={
           <ViewSourceDatasets
+            atlasSourceDatasets={atlasSourceDatasets}
             formManager={formManager}
             isCELLXGENECollection={Boolean(sourceStudy?.cellxgeneCollectionId)}
             pathParameter={pathParameter}
